app/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
email-validator

alembic

psycopg2-binary

PyMySQL


python-dotenv